refactor(workspace): absorb debounce, delete @epicenter/util
@epicenter/util was a single-function package (just `debounce`). It sat in
@epicenter/workspace's dependency closure, which blocks publishing workspace as a
self-contained npm package. Move `debounce` (and its test) into
@epicenter/workspace, export it from the root barrel, repoint every consumer
(workspace internals, @epicenter/filesystem, opensidian, whispering) at the
workspace export, and delete the package.

All four consumers already depend on @epicenter/workspace, so no new dependency
or cycle is introduced. Part of the framework-decomposition prep for jsrepo
distribution (spec 20260614T120000): shrink the publishable surface instead of
committing a junk public package name.
