+++ apps/opensidian/src/lib/state/palette-search-state.svelte.ts
@@ -1,7 +1,7 @@
 import { asFileId } from '@epicenter/filesystem';
 import { createPersistedState } from '@epicenter/svelte';
 import type { CommandPaletteItem } from '@epicenter/ui/command-palette';
-import { debounce } from '@epicenter/util';
+import { debounce } from '@epicenter/workspace';
 import { type } from 'arktype';
 import type { OpensidianBrowser } from 'opensidian/browser';
 import { getFileIcon } from '$lib/utils/file-icons';
