+++ packages/filesystem/package.json
@@ -25,7 +25,6 @@
 	},
 	"dependencies": {
 		"@epicenter/field": "workspace:*",
-		"@epicenter/util": "workspace:*",
 		"@epicenter/workspace": "workspace:*",
 		"@libsql/client": "^0.14.0",
 		"@libsql/client-wasm": "^0.14.0",
