+++ packages/filesystem/src/extensions/sqlite-index/index.ts
@@ -36,8 +36,8 @@
  * @module
  */
 
-import { debounce } from '@epicenter/util';
 import type { Table } from '@epicenter/workspace';
+import { debounce } from '@epicenter/workspace';
 import type { InStatement } from '@libsql/client-wasm';
 import { createClient } from '@libsql/client-wasm';
