--- packages/util/package.json
@@ -1,18 +1,0 @@
-{
-	"name": "@epicenter/util",
-	"description": "Framework-agnostic runtime utilities shared across the Epicenter monorepo",
-	"private": true,
-	"version": "0.0.1",
-	"type": "module",
-	"exports": {
-		".": "./src/index.ts"
-	},
-	"license": "MIT",
-	"scripts": {
-		"typecheck": "tsc --noEmit"
-	},
-	"devDependencies": {
-		"@types/bun": "catalog:",
-		"typescript": "catalog:"
-	}
-}
