--- packages/util/src/index.ts
@@ -1,9 +1,0 @@
-/**
- * `@epicenter/util`: framework-agnostic runtime utilities shared across the
- * Epicenter monorepo. No framework coupling, safe to
- * import from packages, extensions, and apps alike.
- *
- * @module
- */
-
-export { debounce } from './debounce.js';
