--- packages/util/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"types": ["bun"],
-		"noUnusedLocals": true,
-		"noUnusedParameters": true
-	}
-}
