+++ packages/workspace/package.json
@@ -52,7 +52,6 @@
 		"@hono/standard-validator": "catalog:",
 		"@tanstack/ai": "catalog:",
 		"@epicenter/sync": "workspace:*",
-		"@epicenter/util": "workspace:*",
 		"@sindresorhus/slugify": "catalog:",
 		"@types/bun": "catalog:",
 		"arkregex": "catalog:",
