+++ packages/workspace/src/document/attach-encrypted-indexed-db.ts
@@ -50,7 +50,6 @@ import {
 	encryptBytes,
 	type Keyring,
 } from '@epicenter/encryption';
-import { debounce } from '@epicenter/util';
 import * as idb from 'lib0/indexeddb';
 import {
 	defineErrors,
@@ -61,6 +60,7 @@ import { createLogger, type Logger } from 'wellcrafted/logger';
 import { clearDocument } from 'y-indexeddb';
 import type * as Y from 'yjs';
 import { applyUpdateV2, encodeStateAsUpdateV2, transact } from 'yjs';
+import { debounce } from '../shared/debounce.js';
 import type { IndexedDbAttachment } from './attach-indexed-db.js';
 import { deriveWorkspaceKeyring } from './derive-workspace-keyring.js';
